feat: add SavePageModel and createPage init
