+++ packages/iceworks-client/src/pages/Project/components/PagePanel/SavePageModel.module.scss
@@ -1,0 +1,20 @@
+.form {
+  .input {
+    width: 200px;
+    height: 24px;
+    line-height: 24px;
+  }
+
+  .opts {
+    text-align: center;
+    margin-top: 32px;
+
+    .button {
+      margin-right: 12px;
+
+      &:last-child {
+        margin-right: 0;
+      }
+    }
+  }
+}
